Files prepped for decomp
