+++ config/splat.us.stnz0.yaml
@@ -114,6 +114,7 @@ segments:
       - [0x35778, c, gaibon]
       - [0x36DE4, c]
       - [0x374CC, c]
+      - [0x38E0C, c]
       - [0x394D4, c, random]
       - [0x39504, c, update]
       - [0x39800, c, update_stage_entities]
@@ -141,9 +142,5 @@ segments:
       - [0x48ADC, c, e_stage_name] # EntityStageNamePopup
       - [0x49930, c, prim_helpers]
       - [0x4A160, c, lifeupspawn]
-<<<<<<< HEAD
-      - [0x4AA10, sbss]
-=======
       - [0x4AA14, sbss]
->>>>>>> d8db4359 (Decompile NZ0 EntityLifeUpSpawn (#1288))
   - [0x4B780]
